fix: Fix broken `long` conversion

+++ package/cpp/FilamentRecorder.cpp
@@ -9,9 +9,9 @@ namespace margelo {
 
 using namespace facebook;
 
-FilamentRecorder::FilamentRecorder(int width, int height, int fps, long bitRate)
+FilamentRecorder::FilamentRecorder(int width, int height, int fps, double bitRate)
     : HybridObject("FilamentRecorder"), _width(width), _height(height), _fps(fps), _bitRate(bitRate) {
-  Logger::log(TAG, "Creating %zu x %zu @ %zu FPS (%zu bps) FilamentRecorder...", width, height, fps, bitRate);
+  Logger::log(TAG, "Creating %zu x %zu @ %zu FPS (%f bps) FilamentRecorder...", width, height, fps, bitRate);
 }
 
 FilamentRecorder::~FilamentRecorder() {
